pyproject.toml
[project]
name = "entity-resolution-poc"
version = "0.1.0"
description = "Research POC: beating BM25 at people-entity resolution with fine-tuned Matryoshka embeddings at 500M scale"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
    { name = "jayshah5696", email = "jayshah5696@gmail.com" }
]
license = { text = "MIT" }

dependencies = [
    "faker>=24.0",
    "sentence-transformers>=3.3",
    "accelerate>=1.2.0",
    "faiss-cpu>=1.8",
    "rank-bm25>=0.2",
    "datasets>=2.18",
    "transformers>=4.40",
    "torch>=2.2",
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "pydantic>=2.6",
    "rich>=13.7",
    "typer>=0.12",
    "polars>=0.20",
    "pyarrow>=15.0",
    "python-levenshtein>=0.25",
    "tqdm>=4.66",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "wandb>=0.16",
    "pyyaml>=6.0",
    "lancedb>=0.6",
    "modal>=1.3.5",
    "hf-xet>=1.3.2",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ipykernel>=6.29",
    "jupyterlab>=4.1",
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = false
